+++ src/moduler/aktivitet/visning/hjelpekomponenter/forhaandsorientering/LestDatoVisning.tsx
@@ -2,14 +2,14 @@ import { Normaltekst } from 'nav-frontend-typografi';
 import React from 'react';
 
 import { formaterDatoKortManed } from '../../../../../utils';
-import styles from './LestDatoForhaandsorientering.module.less';
+import styles from './LestDatoVisning.module.less';
 
 interface Props {
     hidden: boolean;
     lestDato?: string;
 }
 
-const LestDatoForhaandsorientering = (props: Props) => {
+const LestDatoVisning = (props: Props) => {
     const { hidden, lestDato } = props;
 
     if (hidden) {
@@ -19,4 +19,4 @@ const LestDatoForhaandsorientering = (props: Props) => {
     return <Normaltekst className={styles.lestDato}>Lest {formaterDatoKortManed(lestDato)}</Normaltekst>;
 };
 
-export default LestDatoForhaandsorientering;
+export default LestDatoVisning;
